add `cpprinterface.cpp`
