+++ src/igraph_vector.hpp
